feat(math): separate `PrimeFieldTraits` and define `kIsPrimeField`
Using this, you can determine whether `Field<T>` is `PrimeField<T>` or not.
