+++ tachyon/math/finite_fields/BUILD.bazel
@@ -25,7 +25,7 @@ tachyon_cc_library(
     name = "prime_field_base",
     hdrs = ["prime_field_base.h"],
     deps = [
-        ":prime_field_forward",
+        ":prime_field_traits",
         ":prime_field_util",
         "//tachyon/base:bits",
         "//tachyon/base/strings:string_number_conversions",
@@ -100,6 +100,12 @@ tachyon_cc_library(
 tachyon_cc_library(
     name = "prime_field_forward",
     hdrs = ["prime_field_forward.h"],
+)
+
+tachyon_cc_library(
+    name = "prime_field_traits",
+    hdrs = ["prime_field_traits.h"],
+    deps = [":prime_field_forward"],
 )
 
 tachyon_cc_library(
