+++ tachyon/math/finite_fields/prime_field_base.h
@@ -8,6 +8,7 @@
 #include "tachyon/math/base/field.h"
 #include "tachyon/math/base/gmp/gmp_util.h"
 #include "tachyon/math/finite_fields/prime_field_forward.h"
+#include "tachyon/math/finite_fields/prime_field_traits.h"
 #include "tachyon/math/finite_fields/prime_field_util.h"
 
 namespace tachyon::math {
