+++ tachyon/math/finite_fields/prime_field_forward.h
@@ -15,29 +15,6 @@ class PrimeFieldGpu;
 template <typename Config>
 class PrimeFieldGpuDebug;
 
-template <typename T>
-struct PrimeFieldTraits;
-
-template <typename _Config>
-struct PrimeFieldTraits<PrimeField<_Config>> {
-  using Config = _Config;
-};
-
-template <typename _Config>
-struct PrimeFieldTraits<PrimeFieldGmp<_Config>> {
-  using Config = _Config;
-};
-
-template <typename _Config>
-struct PrimeFieldTraits<PrimeFieldGpu<_Config>> {
-  using Config = _Config;
-};
-
-template <typename _Config>
-struct PrimeFieldTraits<PrimeFieldGpuDebug<_Config>> {
-  using Config = _Config;
-};
-
 }  // namespace tachyon::math
 
 #endif  // TACHYON_MATH_FINITE_FIELDS_PRIME_FIELD_FORWARD_H_
